+++ scripts/code-check.js
@@ -10,7 +10,7 @@ let fs = require('fs-extra')
 let abs = require('path').resolve
 
 // Show message
-alert('Standard JavaScript ' + (env.arg.fix === true ? 'fix' : 'check') + ' ongoing - please wait ...')
+alert('Code ' + (env.arg.fix === true ? 'fix' : 'check') + ' ongoing - please wait ...')
 
 // Define log file
 let logFile = 'code-unconformities.log'
@@ -18,8 +18,7 @@ let logFile = 'code-unconformities.log'
 // Define Standard parameters
 let params = [
   'node',
-  'cmd.js',
-  '>"' + abs(env.proj, logFile) + '"',
+  'node_modules/eslint/bin/eslint.js',
   // Find app.vue, pages/*.vue and pages/sub/*.vue
   '"' + abs(env.app, '**/*.vue') + '"',
   '"' + abs(env.app, '**/*.js') + '"'
@@ -28,8 +27,9 @@ if (!env.installed) {
   params.push('"' + abs(__dirname, '../client/*.js') + '"')
   params.push('"' + abs(__dirname, '../scripts/*.js') + '"')
 }
-params.push('--plugin')
-params.push('html')
+// Print to log file
+params.push('>')
+params.push('"' + abs(env.proj, logFile) + '"')
 if (env.arg.fix === true) {
   params.push('--fix')
 }
@@ -39,10 +39,22 @@ let errorAlert = (env.arg.fix !== true ? 'Code unconformities found.' : 'Some un
                  'Please check "' + logFile + '" for detailed information.\n' +
                  (env.arg.fix !== true ? 'You can run "npm run fix" first for automatic fix.' : '')
 
+// Update eslint config file
+const eslintConfig = env.cfg.eslint
+if (eslintConfig.extends === 'airbnb') {
+  eslintConfig.extends = 'airbnb-base'
+}
+eslintConfig.plugins = ['vue']
+try {
+  fs.writeJsonSync(abs(env.proj, '.eslintrc'), eslintConfig, { spaces: 2 })
+} catch (err) {
+  alert('Failed to update the ESLint configuration file.', 'issue')
+}
+
 // Do the fix
-cmd([env.proj, 'node_modules/standard/bin'], params, function () {
+cmd([env.proj], params, function () {
   fs.removeSync(abs(env.proj, logFile))
-  alert('Standard JavaScript ' + (env.arg.fix === true ? 'fix' : 'check') + ' done without findings.')
+  alert('Code ' + (env.arg.fix === true ? 'fix' : 'check') + ' done without findings.')
 }, function () {
   alert(errorAlert, 'error')
 })