+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Released on 2017-10-26
 
 ### New features
 
-- [x] [#593 - Make it possible to edit root vue component](https://github.com/scriptPilot/app-framework/issues/593)
+- [x] [#593 - Make it possible to edit root vue component](https://github.com/scriptPilot/app-framework/issues/593) ([read the docs](https://github.com/scriptPilot/app-framework/blob/master/docs/extend-vue-root-object.md))
 
 ## Version 1.14.7